pyproject.toml
[project]
name = "marketing-agent"
version = "0.1.0"
description = "Persona Marketing Agent mit Google ADK und Gradio"
requires-python = ">=3.11"

dependencies = [
    "google-adk>=1.20.0",
    "gradio>=6.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["marketing_agent"]